Toegevoegen laatste wijzigenen project.

+++ test.int/app/Http/Controllers/AdminController.php
@@ -6,8 +6,9 @@ use Illuminate\Http\Request;
 use App\User;
 use App\First_period_answer;
 use App\Second_period_answer;
-use App\Third_pepriod_answer;
+use App\Third_period_answer;
 use App\Fourth_period_answer;
+use Excel;
 
 class AdminController extends Controller
 {
@@ -18,17 +19,49 @@ class AdminController extends Controller
     }
 
 
+    /*
+        De manage functie returnt al de Users en eager load de antwoorden.
+        De admin kan hier enkel komen en heeft zo een overzicht over al de gebruikers en hun antwoorden.
+    */
+
     public function manage()
     {
 
     	$users = User::simplePaginate(15);
     	$users->load('First_period_answer', 'Second_period_answer', 'Third_period_answer', 'Fourth_period_answer');
-    	return view('admin.manage', compact('users'));
+        $message = session('message');
+    	return view('admin.manage', compact('users', 'message'));
     }
 
-    public function delete(User $user)
+    /*
+        De delete functie zorgt ervoor dat een gekoze user verwijderd kan worden.
+        hierbij worden ook al de gegeven antwoorden van die user verwijderd.
+    */
+
+    public function delete(User $user, Request $request)
     {
-    	$user->delete();
-    	return redirect('/manage');
+        try{
+
+        $userAnswer1 = First_period_answer::where('user_id', $user->id)->delete();
+        $userAnswer2 = Second_period_answer::where('user_id', $user->id)->delete();
+        $userAnswer3 = Third_period_answer::where('user_id', $user->id)->delete();
+        $userAnswer4 = Fourth_period_answer::where('user_id', $user->id)->delete();
+        $user->delete();
+        $request->session()->flash('message', "Account is succesvol verwijderd.");
+        return redirect('/manage');
+
+        }
+        catch(Exception $e)
+        {
+        $request->session()->flash('message', "Er is iets mis gelopen. We lossen het zo snel mogelijk op.");
+        return redirect('/manage');
+        }
+    	
+
+
+    	
     }
+
+    
 }
+
